* Renamed router file names to use "Router" suffix. * Fix SubjectsRouter since it was querying the wrong table
